+++ sample-project/www/js/index.js
@@ -53,7 +53,6 @@ onDeviceReady: function() {
     document.getElementById("getUnreadCardCountForCategoriesBtn").addEventListener("click", getUnreadCardCountForCategories);
     document.getElementById("getAllNewsFeedCardsBtn").addEventListener("click", getAllNewsFeedCards);
     document.getElementById("getAllContentCardsBtn").addEventListener("click", getContentCardsFromServer);
-    document.getElementById("logContentCardsDisplayedBtn").addEventListener("click", logContentCardsDisplayed);
     document.getElementById("logContentCardAnalyticsBtn").addEventListener("click", logContentCardAnalytics);
     document.getElementById("addAliasBtn").addEventListener("click", addAlias);
     document.getElementById("wipeData").addEventListener("click", wipeData);
@@ -253,10 +252,6 @@ function getAllNewsFeedCards() {
 
 function getContentCardsFromServer() {
     AppboyPlugin.getContentCardsFromServer(customPluginSuccessArrayCallback("test"), customPluginErrorCallback);
-}
-
-function logContentCardsDisplayed() {
-    AppboyPlugin.logContentCardsDisplayed();
 }
 
 function logContentCardAnalytics() {