+++ src/android/AppboyPlugin.java
@@ -72,7 +72,6 @@ public class AppboyPlugin extends CordovaPlugin {
   // Content Card method names
   private static final String GET_CONTENT_CARDS_FROM_SERVER_METHOD = "getContentCardsFromServer";
   private static final String GET_CONTENT_CARDS_FROM_CACHE_METHOD = "getContentCardsFromCache";
-  private static final String LOG_CONTENT_CARDS_DISPLAYED_METHOD = "logContentCardsDisplayed";
   private static final String LOG_CONTENT_CARDS_CLICKED_METHOD = "logContentCardClicked";
   private static final String LOG_CONTENT_CARDS_IMPRESSION_METHOD = "logContentCardImpression";
   private static final String LOG_CONTENT_CARDS_DISMISSED_METHOD = "logContentCardDismissed";
@@ -311,9 +310,6 @@ public class AppboyPlugin extends CordovaPlugin {
       case GET_CONTENT_CARDS_FROM_SERVER_METHOD:
       case GET_CONTENT_CARDS_FROM_CACHE_METHOD:
         return handleContentCardsUpdateGetters(action, callbackContext);
-      case LOG_CONTENT_CARDS_DISPLAYED_METHOD:
-        Braze.getInstance(mApplicationContext).logContentCardsDisplayed();
-        return true;
       case LOG_CONTENT_CARDS_CLICKED_METHOD:
       case LOG_CONTENT_CARDS_DISMISSED_METHOD:
       case LOG_CONTENT_CARDS_IMPRESSION_METHOD: