+++ src/ios/AppboyPlugin.m
@@ -464,10 +464,6 @@
   }
 }
 
-- (void) logContentCardsDisplayed:(CDVInvokedUrlCommand *)command {
-  [[Appboy sharedInstance] logContentCardsDisplayed];
-}
-
 - (void) getContentCardsFromServer:(CDVInvokedUrlCommand *)command {
   NSNotificationCenter *center = [NSNotificationCenter defaultCenter];
   NSOperationQueue *mainQueue = [NSOperationQueue mainQueue];